+++ packages/image/.eslintignore
@@ -2,4 +2,4 @@
 **/coverage/*
 dist
 rnw.js
-src/react-native-safe-area.js
+src/safearea/react-native-safe-area.js
