+++ packages/image/__tests__/modal-shared-with-style.native.js
@@ -13,7 +13,7 @@ import { iterator } from "@times-components/test-utils";
 import Responsive from "@times-components/responsive";
 import { setIsTablet } from "./mocks";
 
-import ModalImage from "../src/modal-image";
+import ModalImage from "../src/modalImage";
 
 // eslint-disable-next-line react/prop-types
 const MockCaption = ({ style: { text, caption, credits, container } }) => (
