+++ packages/image/src/index.js
@@ -1,4 +1,4 @@
 import Image from "./image";
 
-export { default as ModalImage } from "./modal-image";
+export { default as ModalImage } from "./modalImage";
 export default Image;
