+++ packages/image/src/lazyLoadingImage/index.js
@@ -1,0 +1,3 @@
+import LazyLoadingImage from "./lazy-loading-image";
+
+export default LazyLoadingImage;
