+++ packages/image/src/lazyLoadingImage/lazy-loading-image.js
