--- packages/image/src/modal-image.web.js
@@ -1,3 +1,0 @@
-import Image from "./image";
-
-export default Image;
