+++ packages/image/src/modalCaptionContainer/index.js
@@ -1,0 +1,3 @@
+import CaptionContainer from "./modal-caption-container";
+
+export default CaptionContainer;
