+++ packages/image/src/modalCaptionContainer/modal-caption-container.js
