+++ packages/image/src/modalImage/index.js
@@ -1,0 +1,3 @@
+import ModalImage from "./modal-image";
+
+export default ModalImage;
