+++ packages/image/src/modalImage/modal-image.web.js
@@ -1,0 +1,3 @@
+import Image from "../image";
+
+export default Image;
