+++ packages/image/src/placeholder/index.js
@@ -1,0 +1,3 @@
+import PlaceHolder from "./placeholder";
+
+export default PlaceHolder;
