+++ packages/image/src/placeholder/placeholder.js
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { View } from "react-native";
-import styles from "./styles/index";
-import T from "./t";
+import styles from "../styles/index";
+import T from "../t";
 
 function Placeholder({ dimensions }) {
   if (!dimensions) {
