+++ packages/image/src/safearea/index.js
@@ -1,0 +1,3 @@
+import SafeArea from "./react-native-safe-area";
+
+export default SafeArea;
