+++ packages/image/src/safearea/react-native-safe-area.js
