+++ packages/image/src/utils/index.js
@@ -1,0 +1,3 @@
+import appendURL from "./appendToURL";
+
+export default appendURL
